Cleaned up files and consolidated security related classes into secuirty package.
